feat: create useChamados hook and extract filter and fetch logic
