+++ hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useAuth';
+export * from './useChamados';
 export * from './useToast';
 export * from './useUser';
